episode 5  lets get hooked
