--- app.js
@@ -1,1178 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-const data = [
-  {
-    info: {
-      id: "81121",
-      name: "Khichadiwala",
-      cloudinaryImageId: "xp6olehaqooxp5zxocxp",
-      locality: "6th Sector",
-      areaName: "HSR Layout",
-      costForTwo: "₹200 for two",
-      cuisines: ["North Indian"],
-      avgRating: 4.5,
-      veg: true,
-      parentId: "4829",
-      avgRatingString: "4.5",
-      totalRatingsString: "10K+",
-      promoted: true,
-      adTrackingId:
-        "cid=10328751~p=1~eid=0000018c-c8ed-278f-0df2-eadb00890156~srvts=1704178034575~45995",
-      sla: {
-        deliveryTime: 17,
-        lastMileTravel: 2.3,
-        serviceability: "SERVICEABLE",
-        slaString: "17 mins",
-        lastMileTravelString: "2.3 km",
-        iconType: "ICON_TYPE_EMPTY",
-      },
-      availability: {
-        nextCloseTime: "2024-01-02 23:35:00",
-        opened: true,
-      },
-      badges: {
-        imageBadges: [
-          {
-            imageId: "v1695133679/badges/Pure_Veg111.png",
-            description: "pureveg",
-          },
-        ],
-      },
-      isOpen: true,
-      type: "F",
-      badgesV2: {
-        entityBadges: {
-          textBased: {},
-          imageBased: {
-            badgeObject: [
-              {
-                attributes: {
-                  description: "pureveg",
-                  imageId: "v1695133679/badges/Pure_Veg111.png",
-                },
-              },
-            ],
-          },
-          textExtendedBadges: {},
-        },
-      },
-      aggregatedDiscountInfoV3: {
-        header: "50% OFF",
-        subHeader: "UPTO ₹100",
-      },
-      differentiatedUi: {
-        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-        differentiatedUiMediaDetails: {
-          mediaType: "ADS_MEDIA_ENUM_IMAGE",
-          lottie: {},
-          video: {},
-        },
-      },
-      reviewsSummary: {},
-      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-      restaurantOfferPresentationInfo: {},
-    },
-    analytics: {},
-    cta: {
-      link: "swiggy://menu?restaurant_id=81121",
-      text: "RESTAURANT_MENU",
-      type: "DEEPLINK",
-    },
-  },
-  {
-    info: {
-      id: "428",
-      name: "Biryani Pot",
-      cloudinaryImageId: "mdipoyzfzsa7n7igskht",
-      locality: "Maruti Nagar",
-      areaName: "Btm Layout",
-      costForTwo: "₹500 for two",
-      cuisines: ["North Indian", "Biryani"],
-      avgRating: 3.9,
-      parentId: "21798",
-      avgRatingString: "3.9",
-      totalRatingsString: "10K+",
-      sla: {
-        deliveryTime: 18,
-        lastMileTravel: 1.4,
-        serviceability: "SERVICEABLE",
-        slaString: "18 mins",
-        lastMileTravelString: "1.4 km",
-        iconType: "ICON_TYPE_EMPTY",
-      },
-      availability: {
-        nextCloseTime: "2024-01-03 03:00:00",
-        opened: true,
-      },
-      badges: {
-        textExtendedBadges: [
-          {
-            iconId: "guiltfree/GF_Logo_android_3x",
-            shortDescription: "options available",
-            fontColor: "#7E808C",
-          },
-        ],
-      },
-      isOpen: true,
-      type: "F",
-      badgesV2: {
-        entityBadges: {
-          textBased: {},
-          imageBased: {},
-          textExtendedBadges: {
-            badgeObject: [
-              {
-                attributes: {
-                  iconId: "guiltfree/GF_Logo_android_3x",
-                  description: "",
-                  shortDescription: "options available",
-                  fontColor: "#7E808C",
-                },
-              },
-            ],
-          },
-        },
-      },
-      aggregatedDiscountInfoV3: {
-        header: "30% OFF",
-        subHeader: "UPTO ₹75",
-      },
-      differentiatedUi: {
-        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-        differentiatedUiMediaDetails: {
-          mediaType: "ADS_MEDIA_ENUM_IMAGE",
-          lottie: {},
-          video: {},
-        },
-      },
-      reviewsSummary: {},
-      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-      restaurantOfferPresentationInfo: {},
-    },
-    analytics: {},
-    cta: {
-      link: "swiggy://menu?restaurant_id=428",
-      text: "RESTAURANT_MENU",
-      type: "DEEPLINK",
-    },
-  },
-  {
-    info: {
-      id: "664075",
-      name: "Mani's Dum Biryani",
-      cloudinaryImageId: "e6bfe20a8b03779efebeb04eae9a86c6",
-      locality: "2nd Stage",
-      areaName: "BTM Layout",
-      costForTwo: "₹400 for two",
-      cuisines: ["Andhra", "Biryani", "North Indian"],
-      avgRating: 4.4,
-      parentId: "623",
-      avgRatingString: "4.4",
-      totalRatingsString: "1K+",
-      sla: {
-        deliveryTime: 14,
-        lastMileTravel: 1,
-        serviceability: "SERVICEABLE",
-        slaString: "14 mins",
-        lastMileTravelString: "1.0 km",
-        iconType: "ICON_TYPE_EMPTY",
-      },
-      availability: {
-        nextCloseTime: "2024-01-02 23:45:00",
-        opened: true,
-      },
-      badges: {},
-      isOpen: true,
-      aggregatedDiscountInfoV2: {},
-      type: "F",
-      badgesV2: {
-        entityBadges: {
-          textBased: {},
-          imageBased: {},
-          textExtendedBadges: {},
-        },
-      },
-      differentiatedUi: {
-        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-        differentiatedUiMediaDetails: {
-          mediaType: "ADS_MEDIA_ENUM_IMAGE",
-          lottie: {},
-          video: {},
-        },
-      },
-      reviewsSummary: {},
-      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-      restaurantOfferPresentationInfo: {},
-    },
-    analytics: {},
-    cta: {
-      link: "swiggy://menu?restaurant_id=664075",
-      text: "RESTAURANT_MENU",
-      type: "DEEPLINK",
-    },
-  },
-  {
-    info: {
-      id: "201224",
-      name: "Asha tiffins",
-      cloudinaryImageId: "n15vckntsiboiod3gpco",
-      locality: "7th Sector",
-      areaName: "Hsr Layout",
-      costForTwo: "₹200 for two",
-      cuisines: [
-        "South Indian",
-        "North Indian",
-        "Chinese",
-        "Desserts",
-        "Beverages",
-      ],
-      avgRating: 4.4,
-      veg: true,
-      parentId: "236243",
-      avgRatingString: "4.4",
-      totalRatingsString: "10K+",
-      sla: {
-        deliveryTime: 17,
-        lastMileTravel: 2.5,
-        serviceability: "SERVICEABLE",
-        slaString: "17 mins",
-        lastMileTravelString: "2.5 km",
-        iconType: "ICON_TYPE_EMPTY",
-      },
-      availability: {
-        nextCloseTime: "2024-01-02 15:30:00",
-        opened: true,
-      },
-      badges: {
-        imageBadges: [
-          {
-            imageId: "v1695133679/badges/Pure_Veg111.png",
-            description: "pureveg",
-          },
-        ],
-      },
-      isOpen: true,
-      type: "F",
-      badgesV2: {
-        entityBadges: {
-          textBased: {},
-          imageBased: {
-            badgeObject: [
-              {
-                attributes: {
-                  imageId: "v1695133679/badges/Pure_Veg111.png",
-                  description: "pureveg",
-                },
-              },
-            ],
-          },
-          textExtendedBadges: {},
-        },
-      },
-      aggregatedDiscountInfoV3: {
-        header: "50% OFF",
-        subHeader: "UPTO ₹100",
-      },
-      differentiatedUi: {
-        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-        differentiatedUiMediaDetails: {
-          mediaType: "ADS_MEDIA_ENUM_IMAGE",
-          lottie: {},
-          video: {},
-        },
-      },
-      reviewsSummary: {},
-      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-      restaurantOfferPresentationInfo: {},
-    },
-    analytics: {},
-    cta: {
-      link: "swiggy://menu?restaurant_id=201224",
-      text: "RESTAURANT_MENU",
-      type: "DEEPLINK",
-    },
-  },
-  {
-    info: {
-      id: "326590",
-      name: "O4H - Order For Health",
-      cloudinaryImageId: "im7blwmj0j1xbjw1ysi3",
-      locality: "2nd Stage",
-      areaName: "BTM Layout",
-      costForTwo: "₹400 for two",
-      cuisines: [
-        "Healthy Food",
-        "Salads",
-        "Keto",
-        "Mediterranean",
-        "Beverages",
-        "Desserts",
-        "Continental",
-      ],
-      avgRating: 4.6,
-      parentId: "11615",
-      avgRatingString: "4.6",
-      totalRatingsString: "5K+",
-      promoted: true,
-      adTrackingId:
-        "cid=10240154~p=3~eid=0000018c-c8ed-278f-0df2-eadd0089036e~srvts=1704178034575~45995",
-      sla: {
-        deliveryTime: 18,
-        lastMileTravel: 1.6,
-        serviceability: "SERVICEABLE",
-        slaString: "18 mins",
-        lastMileTravelString: "1.6 km",
-        iconType: "ICON_TYPE_EMPTY",
-      },
-      availability: {
-        nextCloseTime: "2024-01-02 23:00:00",
-        opened: true,
-      },
-      badges: {
-        textExtendedBadges: [
-          {
-            iconId: "guiltfree/GF_Logo_android_3x",
-            shortDescription: "brand",
-            fontColor: "#7E808C",
-          },
-        ],
-      },
-      isOpen: true,
-      type: "F",
-      badgesV2: {
-        entityBadges: {
-          textBased: {},
-          imageBased: {},
-          textExtendedBadges: {
-            badgeObject: [
-              {
-                attributes: {
-                  iconId: "guiltfree/GF_Logo_android_3x",
-                  description: "",
-                  shortDescription: "brand",
-                  fontColor: "#7E808C",
-                },
-              },
-            ],
-          },
-        },
-      },
-      aggregatedDiscountInfoV3: {
-        header: "10% OFF",
-        subHeader: "ABOVE ₹800",
-        discountTag: "FLAT DEAL",
-      },
-      differentiatedUi: {
-        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-        differentiatedUiMediaDetails: {
-          mediaType: "ADS_MEDIA_ENUM_IMAGE",
-          lottie: {},
-          video: {},
-        },
-      },
-      reviewsSummary: {},
-      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-      restaurantOfferPresentationInfo: {},
-    },
-    analytics: {},
-    cta: {
-      link: "swiggy://menu?restaurant_id=326590",
-      text: "RESTAURANT_MENU",
-      type: "DEEPLINK",
-    },
-  },
-  {
-    info: {
-      id: "334475",
-      name: "KFC",
-      cloudinaryImageId: "f01666ac73626461d7455d9c24005cd4",
-      locality: "2nd Stage",
-      areaName: "BTM Layout",
-      costForTwo: "₹400 for two",
-      cuisines: ["Burgers", "Biryani", "American", "Snacks", "Fast Food"],
-      avgRating: 4,
-      parentId: "547",
-      avgRatingString: "4.0",
-      totalRatingsString: "5K+",
-      sla: {
-        deliveryTime: 15,
-        lastMileTravel: 0.9,
-        serviceability: "SERVICEABLE",
-        slaString: "15 mins",
-        lastMileTravelString: "0.9 km",
-        iconType: "ICON_TYPE_EMPTY",
-      },
-      availability: {
-        nextCloseTime: "2024-01-03 02:00:00",
-        opened: true,
-      },
-      badges: {
-        textExtendedBadges: [
-          {
-            iconId: "v1702882113/Ratnesh_Badges/20x20_px_kids__20px.png",
-            shortDescription: "Kids Favourite options",
-            fontColor: "#7E808C",
-          },
-        ],
-      },
-      isOpen: true,
-      type: "F",
-      badgesV2: {
-        entityBadges: {
-          textBased: {},
-          imageBased: {},
-          textExtendedBadges: {
-            badgeObject: [
-              {
-                attributes: {
-                  fontColor: "#7E808C",
-                  iconId: "v1702882113/Ratnesh_Badges/20x20_px_kids__20px.png",
-                  description: "",
-                  shortDescription: "Kids Favourite options",
-                },
-              },
-            ],
-          },
-        },
-      },
-      aggregatedDiscountInfoV3: {
-        header: "20% OFF",
-        subHeader: "UPTO ₹50",
-      },
-      differentiatedUi: {
-        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-        differentiatedUiMediaDetails: {
-          mediaType: "ADS_MEDIA_ENUM_IMAGE",
-          lottie: {},
-          video: {},
-        },
-      },
-      reviewsSummary: {},
-      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-      restaurantOfferPresentationInfo: {},
-    },
-    analytics: {},
-    cta: {
-      link: "swiggy://menu?restaurant_id=334475",
-      text: "RESTAURANT_MENU",
-      type: "DEEPLINK",
-    },
-  },
-  {
-    info: {
-      id: "292770",
-      name: "Charminar Biryani House",
-      cloudinaryImageId: "l8z0w7mrwwlkky0tumgv",
-      locality: "1st  Stage",
-      areaName: "BTM Layout",
-      costForTwo: "₹200 for two",
-      cuisines: ["Biryani", "North Indian"],
-      avgRating: 4.1,
-      parentId: "58180",
-      avgRatingString: "4.1",
-      totalRatingsString: "1K+",
-      sla: {
-        deliveryTime: 16,
-        lastMileTravel: 0.9,
-        serviceability: "SERVICEABLE",
-        slaString: "16 mins",
-        lastMileTravelString: "0.9 km",
-        iconType: "ICON_TYPE_EMPTY",
-      },
-      availability: {
-        nextCloseTime: "2024-01-03 03:00:00",
-        opened: true,
-      },
-      badges: {
-        textExtendedBadges: [
-          {
-            iconId: "guiltfree/GF_Logo_android_3x",
-            shortDescription: "options available",
-            fontColor: "#7E808C",
-          },
-        ],
-      },
-      isOpen: true,
-      type: "F",
-      badgesV2: {
-        entityBadges: {
-          textBased: {},
-          imageBased: {},
-          textExtendedBadges: {
-            badgeObject: [
-              {
-                attributes: {
-                  fontColor: "#7E808C",
-                  iconId: "guiltfree/GF_Logo_android_3x",
-                  description: "",
-                  shortDescription: "options available",
-                },
-              },
-            ],
-          },
-        },
-      },
-      aggregatedDiscountInfoV3: {
-        header: "30% OFF",
-        discountTag: "FLAT DEAL",
-      },
-      differentiatedUi: {
-        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-        differentiatedUiMediaDetails: {
-          mediaType: "ADS_MEDIA_ENUM_IMAGE",
-          lottie: {},
-          video: {},
-        },
-      },
-      reviewsSummary: {},
-      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-      restaurantOfferPresentationInfo: {},
-    },
-    analytics: {},
-    cta: {
-      link: "swiggy://menu?restaurant_id=292770",
-      text: "RESTAURANT_MENU",
-      type: "DEEPLINK",
-    },
-  },
-  {
-    info: {
-      id: "211192",
-      name: "La Pino'z Pizza",
-      cloudinaryImageId: "lczhp9lptdzbqn09nfns",
-      locality: "2nd Stage",
-      areaName: "BTM Layout",
-      costForTwo: "₹250 for two",
-      cuisines: ["Pizzas", "Pastas", "Italian", "Desserts", "Beverages"],
-      avgRating: 4,
-      parentId: "4961",
-      avgRatingString: "4.0",
-      totalRatingsString: "5K+",
-      sla: {
-        deliveryTime: 18,
-        lastMileTravel: 0.9,
-        serviceability: "SERVICEABLE",
-        slaString: "18 mins",
-        lastMileTravelString: "0.9 km",
-        iconType: "ICON_TYPE_EMPTY",
-      },
-      availability: {
-        nextCloseTime: "2024-01-03 03:00:00",
-        opened: true,
-      },
-      badges: {},
-      isOpen: true,
-      type: "F",
-      badgesV2: {
-        entityBadges: {
-          imageBased: {},
-          textExtendedBadges: {},
-          textBased: {},
-        },
-      },
-      aggregatedDiscountInfoV3: {
-        header: "50% OFF",
-        subHeader: "UPTO ₹100",
-      },
-      differentiatedUi: {
-        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-        differentiatedUiMediaDetails: {
-          mediaType: "ADS_MEDIA_ENUM_IMAGE",
-          lottie: {},
-          video: {},
-        },
-      },
-      reviewsSummary: {},
-      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-      restaurantOfferPresentationInfo: {},
-    },
-    analytics: {},
-    cta: {
-      link: "swiggy://menu?restaurant_id=211192",
-      text: "RESTAURANT_MENU",
-      type: "DEEPLINK",
-    },
-  },
-  {
-    info: {
-      id: "307",
-      name: "A2B - Adyar Ananda Bhavan",
-      cloudinaryImageId: "gxe4mn4ubw70yx3flyar",
-      locality: "1st Stage",
-      areaName: "Btm Layout",
-      costForTwo: "₹300 for two",
-      cuisines: ["South Indian", "North Indian", "Sweets", "Chinese"],
-      avgRating: 4.4,
-      veg: true,
-      parentId: "22",
-      avgRatingString: "4.4",
-      totalRatingsString: "10K+",
-      promoted: true,
-      adTrackingId:
-        "cid=10373920~p=4~eid=0000018c-c8ed-278f-0df2-eade0089045c~srvts=1704178034575~45995",
-      sla: {
-        deliveryTime: 15,
-        lastMileTravel: 0.7,
-        serviceability: "SERVICEABLE",
-        slaString: "15 mins",
-        lastMileTravelString: "0.7 km",
-        iconType: "ICON_TYPE_EMPTY",
-      },
-      availability: {
-        nextCloseTime: "2024-01-02 22:30:00",
-        opened: true,
-      },
-      badges: {
-        imageBadges: [
-          {
-            imageId: "v1695133679/badges/Pure_Veg111.png",
-            description: "pureveg",
-          },
-        ],
-      },
-      isOpen: true,
-      type: "F",
-      badgesV2: {
-        entityBadges: {
-          textBased: {},
-          imageBased: {
-            badgeObject: [
-              {
-                attributes: {
-                  imageId: "v1695133679/badges/Pure_Veg111.png",
-                  description: "pureveg",
-                },
-              },
-            ],
-          },
-          textExtendedBadges: {},
-        },
-      },
-      aggregatedDiscountInfoV3: {
-        header: "20% OFF",
-        subHeader: "UPTO ₹50",
-      },
-      differentiatedUi: {
-        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-        differentiatedUiMediaDetails: {
-          mediaType: "ADS_MEDIA_ENUM_IMAGE",
-          lottie: {},
-          video: {},
-        },
-      },
-      reviewsSummary: {},
-      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-      restaurantOfferPresentationInfo: {},
-    },
-    analytics: {},
-    cta: {
-      link: "swiggy://menu?restaurant_id=307",
-      text: "RESTAURANT_MENU",
-      type: "DEEPLINK",
-    },
-  },
-  {
-    info: {
-      id: "92389",
-      name: "Muthashy's",
-      cloudinaryImageId: "fifv8bhm4mzqausmynfa",
-      locality: "Maruti Nagar",
-      areaName: "Btm Layout",
-      costForTwo: "₹300 for two",
-      cuisines: ["South Indian", "Kerala", "Seafood", "North Indian"],
-      avgRating: 4.3,
-      parentId: "21429",
-      avgRatingString: "4.3",
-      totalRatingsString: "10K+",
-      sla: {
-        deliveryTime: 14,
-        lastMileTravel: 0.8,
-        serviceability: "SERVICEABLE",
-        slaString: "14 mins",
-        lastMileTravelString: "0.8 km",
-        iconType: "ICON_TYPE_EMPTY",
-      },
-      availability: {
-        nextCloseTime: "2024-01-02 23:00:00",
-        opened: true,
-      },
-      badges: {},
-      isOpen: true,
-      aggregatedDiscountInfoV2: {},
-      type: "F",
-      badgesV2: {
-        entityBadges: {
-          imageBased: {},
-          textExtendedBadges: {},
-          textBased: {},
-        },
-      },
-      differentiatedUi: {
-        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-        differentiatedUiMediaDetails: {
-          mediaType: "ADS_MEDIA_ENUM_IMAGE",
-          lottie: {},
-          video: {},
-        },
-      },
-      reviewsSummary: {},
-      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-      restaurantOfferPresentationInfo: {},
-    },
-    analytics: {},
-    cta: {
-      link: "swiggy://menu?restaurant_id=92389",
-      text: "RESTAURANT_MENU",
-      type: "DEEPLINK",
-    },
-  },
-  {
-    info: {
-      id: "23683",
-      name: "McDonald's",
-      cloudinaryImageId: "bb7ae131544c7d37e10fc5faf76f09d6",
-      locality: "2nd Stage",
-      areaName: "Btm Layout",
-      costForTwo: "₹400 for two",
-      cuisines: ["Burgers", "Beverages", "Cafe", "Desserts"],
-      avgRating: 4.2,
-      parentId: "630",
-      avgRatingString: "4.2",
-      totalRatingsString: "10K+",
-      sla: {
-        deliveryTime: 15,
-        lastMileTravel: 1.2,
-        serviceability: "SERVICEABLE",
-        slaString: "15 mins",
-        lastMileTravelString: "1.2 km",
-        iconType: "ICON_TYPE_EMPTY",
-      },
-      availability: {
-        nextCloseTime: "2024-01-03 03:45:00",
-        opened: true,
-      },
-      badges: {
-        textExtendedBadges: [
-          {
-            iconId: "guiltfree/GF_Logo_android_3x",
-            shortDescription: "options available",
-            fontColor: "#7E808C",
-          },
-        ],
-      },
-      isOpen: true,
-      aggregatedDiscountInfoV2: {},
-      type: "F",
-      badgesV2: {
-        entityBadges: {
-          imageBased: {},
-          textExtendedBadges: {
-            badgeObject: [
-              {
-                attributes: {
-                  shortDescription: "options available",
-                  fontColor: "#7E808C",
-                  iconId: "guiltfree/GF_Logo_android_3x",
-                  description: "",
-                },
-              },
-            ],
-          },
-          textBased: {},
-        },
-      },
-      differentiatedUi: {
-        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-        differentiatedUiMediaDetails: {
-          mediaType: "ADS_MEDIA_ENUM_IMAGE",
-          lottie: {},
-          video: {},
-        },
-      },
-      reviewsSummary: {},
-      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-      restaurantOfferPresentationInfo: {},
-    },
-    analytics: {},
-    cta: {
-      link: "swiggy://menu?restaurant_id=23683",
-      text: "RESTAURANT_MENU",
-      type: "DEEPLINK",
-    },
-  },
-  {
-    info: {
-      id: "115918",
-      name: "Oottupura",
-      cloudinaryImageId: "b9bfmzf0fxsobu2sxfvd",
-      locality: "Maruthi Nagar",
-      areaName: "Btm Layout",
-      costForTwo: "₹268 for two",
-      cuisines: ["Kerala", "South Indian"],
-      avgRating: 4.2,
-      parentId: "7485",
-      avgRatingString: "4.2",
-      totalRatingsString: "10K+",
-      sla: {
-        deliveryTime: 18,
-        lastMileTravel: 1.6,
-        serviceability: "SERVICEABLE",
-        slaString: "18 mins",
-        lastMileTravelString: "1.6 km",
-        iconType: "ICON_TYPE_EMPTY",
-      },
-      availability: {
-        nextCloseTime: "2024-01-02 22:30:00",
-        opened: true,
-      },
-      badges: {},
-      isOpen: true,
-      type: "F",
-      badgesV2: {
-        entityBadges: {
-          textBased: {},
-          imageBased: {},
-          textExtendedBadges: {},
-        },
-      },
-      aggregatedDiscountInfoV3: {
-        header: "10% OFF",
-        subHeader: "UPTO ₹40",
-      },
-      differentiatedUi: {
-        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-        differentiatedUiMediaDetails: {
-          mediaType: "ADS_MEDIA_ENUM_IMAGE",
-          lottie: {},
-          video: {},
-        },
-      },
-      reviewsSummary: {},
-      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-      restaurantOfferPresentationInfo: {},
-    },
-    analytics: {},
-    cta: {
-      link: "swiggy://menu?restaurant_id=115918",
-      text: "RESTAURANT_MENU",
-      type: "DEEPLINK",
-    },
-  },
-  {
-    info: {
-      id: "592285",
-      name: "Shyamji's Chole Bhature",
-      cloudinaryImageId: "ugxnshnc5dctksihttvq",
-      locality: "5th Block",
-      areaName: "Koramangala",
-      costForTwo: "₹400 for two",
-      cuisines: ["North Indian", "Beverages", "Snacks"],
-      avgRating: 4.2,
-      veg: true,
-      parentId: "255424",
-      avgRatingString: "4.2",
-      totalRatingsString: "1K+",
-      promoted: true,
-      adTrackingId:
-        "cid=10361005~p=5~eid=0000018c-c8ed-278f-0df2-eadf00890510~srvts=1704178034575~45995",
-      sla: {
-        deliveryTime: 18,
-        lastMileTravel: 3,
-        serviceability: "SERVICEABLE",
-        slaString: "18 mins",
-        lastMileTravelString: "3.0 km",
-        iconType: "ICON_TYPE_EMPTY",
-      },
-      availability: {
-        nextCloseTime: "2024-01-02 20:55:00",
-        opened: true,
-      },
-      badges: {},
-      isOpen: true,
-      type: "F",
-      badgesV2: {
-        entityBadges: {
-          textBased: {},
-          imageBased: {},
-          textExtendedBadges: {},
-        },
-      },
-      aggregatedDiscountInfoV3: {
-        header: "10% OFF",
-        subHeader: "UPTO ₹40",
-      },
-      differentiatedUi: {
-        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-        differentiatedUiMediaDetails: {
-          mediaType: "ADS_MEDIA_ENUM_IMAGE",
-          lottie: {},
-          video: {},
-        },
-      },
-      reviewsSummary: {},
-      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-      restaurantOfferPresentationInfo: {},
-    },
-    analytics: {},
-    cta: {
-      link: "swiggy://menu?restaurant_id=592285",
-      text: "RESTAURANT_MENU",
-      type: "DEEPLINK",
-    },
-  },
-  {
-    info: {
-      id: "379511",
-      name: "Bismillah Hotel",
-      cloudinaryImageId: "vygm7kq9vii6avaeeyyx",
-      locality: "2nd Stage",
-      areaName: "BTM Layout",
-      costForTwo: "₹200 for two",
-      cuisines: ["Chinese"],
-      avgRating: 4,
-      parentId: "21396",
-      avgRatingString: "4.0",
-      totalRatingsString: "100+",
-      sla: {
-        deliveryTime: 19,
-        lastMileTravel: 2.2,
-        serviceability: "SERVICEABLE",
-        slaString: "19 mins",
-        lastMileTravelString: "2.2 km",
-        iconType: "ICON_TYPE_EMPTY",
-      },
-      availability: {
-        nextCloseTime: "2024-01-03 01:15:00",
-        opened: true,
-      },
-      badges: {},
-      isOpen: true,
-      type: "F",
-      badgesV2: {
-        entityBadges: {
-          textExtendedBadges: {},
-          textBased: {},
-          imageBased: {},
-        },
-      },
-      aggregatedDiscountInfoV3: {
-        header: "20% OFF",
-        subHeader: "UPTO ₹50",
-      },
-      differentiatedUi: {
-        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-        differentiatedUiMediaDetails: {
-          mediaType: "ADS_MEDIA_ENUM_IMAGE",
-          lottie: {},
-          video: {},
-        },
-      },
-      reviewsSummary: {},
-      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-      restaurantOfferPresentationInfo: {},
-    },
-    analytics: {},
-    cta: {
-      link: "swiggy://menu?restaurant_id=379511",
-      text: "RESTAURANT_MENU",
-      type: "DEEPLINK",
-    },
-  },
-  {
-    info: {
-      id: "728619",
-      name: "Aromas of Biryani",
-      cloudinaryImageId: "a7f6cc7fac3c851fc697b9c2443b22fa",
-      locality: "6th sector",
-      areaName: "HSR",
-      costForTwo: "₹500 for two",
-      cuisines: ["Biryani", "North Indian", "Tandoor", "Kebabs", "Beverages"],
-      avgRating: 4.1,
-      parentId: "36155",
-      avgRatingString: "4.1",
-      totalRatingsString: "1K+",
-      sla: {
-        deliveryTime: 17,
-        lastMileTravel: 2.6,
-        serviceability: "SERVICEABLE",
-        slaString: "17 mins",
-        lastMileTravelString: "2.6 km",
-        iconType: "ICON_TYPE_EMPTY",
-      },
-      availability: {
-        nextCloseTime: "2024-01-02 16:00:00",
-        opened: true,
-      },
-      badges: {},
-      isOpen: true,
-      type: "F",
-      badgesV2: {
-        entityBadges: {
-          textBased: {},
-          imageBased: {},
-          textExtendedBadges: {},
-        },
-      },
-      aggregatedDiscountInfoV3: {
-        header: "₹125 OFF",
-        subHeader: "ABOVE ₹749",
-        discountTag: "FLAT DEAL",
-      },
-      differentiatedUi: {
-        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-        differentiatedUiMediaDetails: {
-          mediaType: "ADS_MEDIA_ENUM_IMAGE",
-          lottie: {},
-          video: {},
-        },
-      },
-      reviewsSummary: {},
-      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-      restaurantOfferPresentationInfo: {},
-    },
-    analytics: {},
-    cta: {
-      link: "swiggy://menu?restaurant_id=728619",
-      text: "RESTAURANT_MENU",
-      type: "DEEPLINK",
-    },
-  },
-  {
-    info: {
-      id: "422863",
-      name: "Jose Mess",
-      cloudinaryImageId: "lltunkievzpxrrcarfmy",
-      locality: "Maruti Nagar",
-      areaName: "Btm Layout",
-      costForTwo: "₹150 for two",
-      cuisines: ["South Indian"],
-      avgRating: 4.3,
-      parentId: "20129",
-      avgRatingString: "4.3",
-      totalRatingsString: "5K+",
-      sla: {
-        deliveryTime: 17,
-        lastMileTravel: 1.4,
-        serviceability: "SERVICEABLE",
-        slaString: "17 mins",
-        lastMileTravelString: "1.4 km",
-        iconType: "ICON_TYPE_EMPTY",
-      },
-      availability: {
-        nextCloseTime: "2024-01-02 23:00:00",
-        opened: true,
-      },
-      badges: {
-        textExtendedBadges: [
-          {
-            iconId: "guiltfree/GF_Logo_android_3x",
-            shortDescription: "options available",
-            fontColor: "#7E808C",
-          },
-        ],
-      },
-      isOpen: true,
-      aggregatedDiscountInfoV2: {},
-      type: "F",
-      badgesV2: {
-        entityBadges: {
-          textExtendedBadges: {
-            badgeObject: [
-              {
-                attributes: {
-                  iconId: "guiltfree/GF_Logo_android_3x",
-                  description: "",
-                  shortDescription: "options available",
-                  fontColor: "#7E808C",
-                },
-              },
-            ],
-          },
-          textBased: {},
-          imageBased: {},
-        },
-      },
-      differentiatedUi: {
-        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-        differentiatedUiMediaDetails: {
-          mediaType: "ADS_MEDIA_ENUM_IMAGE",
-          lottie: {},
-          video: {},
-        },
-      },
-      reviewsSummary: {},
-      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-      restaurantOfferPresentationInfo: {},
-    },
-    analytics: {},
-    cta: {
-      link: "swiggy://menu?restaurant_id=422863",
-      text: "RESTAURANT_MENU",
-      type: "DEEPLINK",
-    },
-  },
-];
-
-function Header() {
-  return (
-    <>
-      <div className="container">
-        <div className="logo">
-          <img src="./Image-na.png" alt="logo" />
-        </div>
-        <div className="navbar">
-          <ul>
-            <li>Home</li>
-            <li>About</li>
-            <li>Contact Us</li>
-            <li>Cart</li>
-          </ul>
-        </div>
-      </div>
-    </>
-  );
-}
-
-function RestaurantCard(props) {
-  console.log(props.resdata.info,"props");
-  const {resdata} = props
-  const {name,cuisines,areaName,cloudinaryImageId} = resdata.info
-  return (
-    
-    <div className="res-card">
-      <img
-      src={"https://media-assets.swiggy.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_660/"+ cloudinaryImageId
-}
-        alt="meghana"
-        className="res-image"
-      />
-
-      <h2>{name}</h2>
-      <h4> {cuisines.join(", ")} </h4>
-      <h4>{areaName}</h4>
-    </div>
-  );
-}
-function Body() {
-  return (
-    <div>
-      <div className="search">Search</div>
-      <div className="res-container">
-        {data.map((restaurant) => <RestaurantCard  resdata={restaurant} key={restaurant.info.id}/>)}
-      
-
-      </div>
-    </div>
-  );
-}
-function App() {
-  return (
-    <div>
-      <Header />
-      <Body />
-    </div>
-  );
-}
-
-const root = ReactDOM.createRoot(document.getElementById("root"));
-root.render(<App />);
