+++ index.html
@@ -10,7 +10,7 @@
     <div id="root">
         <h1>Hello Akshay</h1>
     </div>
-    <script type="module" src="./app.js">
+    <script type="module" src="./src/app.js">
    
     </script>
   </body>
