+++ src/utils/constant.js
@@ -1,0 +1,3 @@
+
+export const IMAGE_URL = "https://media-assets.swiggy.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_660/"
+
